# More accurate broadcast micro-benchmark