Before
```
       Method | Connections | Protocol |         Mean |      Error |     StdDev |      Op/s |  Gen 0 | Allocated |
------------- |------------ |--------- |-------------:|-----------:|-----------:|----------:|-------:|----------:|
 SendAsyncAll |           1 |     json |     3.696 us |  0.0840 us |  0.2462 us | 270,538.5 | 0.0992 |   6.34 KB |
 SendAsyncAll |           1 |  msgpack |     2.837 us |  0.0550 us |  0.0789 us | 352,538.5 | 0.0229 |   1.02 KB |
 SendAsyncAll |          10 |     json |    25.076 us |  0.4678 us |  0.4376 us |  39,879.5 | 0.0610 |   6.91 KB |
 SendAsyncAll |          10 |  msgpack |    23.572 us |  0.2870 us |  0.2685 us |  42,422.6 |      - |    1.6 KB |
 SendAsyncAll |        1000 |     json | 2,158.075 us | 31.1448 us | 29.1329 us |     463.4 |      - |  68.78 KB |
 SendAsyncAll |        1000 |  msgpack | 2,155.878 us | 40.9439 us | 40.2124 us |     463.8 |      - |  63.48 KB |
```

After
```
       Method | Connections | Protocol |       Mean |     Error |     StdDev |      Op/s |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
------------- |------------ |--------- |-----------:|----------:|-----------:|----------:|-------:|-------:|-------:|----------:|
 SendAsyncAll |           1 |     json |   1.934 us | 0.0386 us |  0.0839 us | 516,948.5 | 0.0763 | 0.0057 | 0.0019 |    6479 B |
 SendAsyncAll |           1 |  msgpack |   1.079 us | 0.0214 us |  0.0278 us | 927,058.1 | 0.0114 | 0.0019 |      - |    1014 B |
 SendAsyncAll |          10 |     json |   7.139 us | 0.1392 us |  0.1905 us | 140,066.4 | 0.0153 | 0.0076 |      - |    6904 B |
 SendAsyncAll |          10 |  msgpack |   4.389 us | 0.1141 us |  0.3348 us | 227,861.8 | 0.0076 |      - |      - |    1139 B |
 SendAsyncAll |        1000 |     json | 450.210 us | 9.1506 us | 19.3018 us |   2,221.2 |      - |      - |      - |   52882 B |
 SendAsyncAll |        1000 |  msgpack | 430.684 us | 8.5807 us | 18.2861 us |   2,321.9 |      - |      - |      - |   14290 B |
```

Found some of that missing RPS @davidfowl 
I'm probably doing something unnecessary with pipes in the cleanup